requirements.txt
Chameleon==2.13-1
EmpProj==0.0
HRProj==0.0
Jinja2==2.7.1
Mako==0.9.0
MarkupSafe==0.18
PasteDeploy==1.5.0
Pygments==1.6
SQLAlchemy==0.8.3
SQLAlchemy-Utils==0.22.1
WTForms==1.0.5
WTForms-Components==0.8.3
WebHelpers2==2.0b5
WebOb==1.2.3
WebTest==2.0.10
beautifulsoup4==4.3.2
colander==1.0b1
decorator==3.4.0
deform==2.0a2
docutils==0.11
jinja==0.0
nose==1.3.0
paginate==0.4.1
peppercorn==0.4
psycopg2==2.5.1
pyramid==1.4.5
pyramid-chameleon==0.1
pyramid-debugtoolbar==1.0.9
pyramid-jinja2==1.9
pyramid-mako==0.3.1
pyramid-tm==0.7
repoze.lru==0.6
six==1.4.1
toolz==0.5.0
transaction==1.4.1
translationstring==1.1
tutorial==0.0.0
uWSGI==2.0
validators==0.5.0
venusian==1.0a8
waitress==0.8.7
zope.deprecation==4.0.2
zope.interface==4.0.5
zope.sqlalchemy==0.7.3
